requirements.txt
aws-cdk.cloud-assembly-schema
aws-cdk.cx-api
aws-cdk-lib
aws-cdk.region-info
